redid get_hash(), implemnted rehash(), add(), to_string(), get_size(), contains(), bucket_list_size()
to_string(): we dont get the same as the expected output we can we have the same ?? hashing and all that.
the add(): is dependent on the rehash().
get_size(): is basicly the self.size .
still have to implement the remvoe(), max_bucket(), zero_bucket_ratio():
but i sketched them out in comments so i dont forget under each function.